ahora no se puede eliminar un cliente con pedidos asociados

+++ routes/clientes.js
@@ -4,9 +4,11 @@ const clienteService = require('../services/cliente');
 const personaService = require('../services/persona');
 const { generatePagination } = require('../services/utils');
 const paginate = require('../middlewares/paginate');
+const { sequelize } = require('../models');
 
 //lista todos los clientes
 router.get('/', async (req, res) => {
+  const { success } = req.query;
   try {
     const clientes = await clienteService.getAllClientes(
       {},
@@ -15,7 +17,18 @@ router.get('/', async (req, res) => {
         nest: true,
       }
     );
-    res.render('clientes/cliente', { results: clientes, req });
+    let mensajeExito;
+    let mensajeError;
+    if (success === 'create') {
+      mensajeExito = 'Cliente agregado con exito';
+    }
+    if (success === 'edit') {
+      mensajeExito = 'Cliente editado con exito';
+    }
+    if (success === 'delete') {
+      mensajeExito = 'Cliente eliminado con exito';
+    }
+    res.render('clientes/cliente', { results: clientes, req, success: mensajeExito });
   } catch (error) {
     res.redirect('/404');
   }
@@ -54,12 +67,16 @@ router.get('/editar/:id', async (req, res) => {
   }
 });
 
-
 router.get('/eliminar/:id', async (req, res) => {
   const { id } = req.params;
+  const { error } = req.query;
   try {
     const cliente = await clienteService.getCliente(id, { raw: true, nest: true });
-    res.render('clientes/eliminar', { cliente, req });
+    let mensajeError = '';
+    if (error) {
+      mensajeError = 'No se puede eliminar el cliente ya que tiene pedidos asociados';
+    }
+    res.render('clientes/eliminar', { cliente, req, errores: mensajeError });
   } catch (error) {
     res.redirect('/404');
   }
@@ -92,7 +109,7 @@ router.post('/', async (req, res) => {
       telefono
     );
     const cliente = await clienteService.createClienteExiste(tipoCliente, persona.id);
-    res.redirect('/clientes');
+    res.redirect('/clientes?success=create');
   } catch (error) {
     try {
       const persona = await personaService.getPersonaArgs({ identificacion });
@@ -129,8 +146,13 @@ router.put('/', async (req, res) => {
 router.delete('/', async (req, res) => {
   const { id } = req.body;
   try {
+    const cliente = await clienteService.getCliente(id);
+    const pedidos = await sequelize.models.Pedido.findAll({ where: { PersonaId: cliente.Persona.id } });
+    if (pedidos.length) {
+      return res.redirect(`clientes/eliminar/${id}?error=1`);
+    }
     await clienteService.deleteCliente(id);
-    res.redirect('/clientes');
+    return res.redirect('/clientes');
   } catch (error) {
     res.redirect('/404');
   }
